# Rewrite Symbol#to_sym and #intern in Ruby