## Benchmark
### VM
It doesn't slow down the interpreter:

```
$ benchmark-driver -v --chruby 'before;after' ':sym.to_sym'
before: ruby 3.2.0dev (2022-11-15T08:34:00Z master 1a9e87fe3a) [x86_64-linux]
after: ruby 3.2.0dev (2022-11-15T08:38:30Z builtin-sym-to-sym 43fb0e94cb) [x86_64-linux]
Warming up --------------------------------------
         :sym.to_sym    42.107M i/s -     42.341M times in 1.005555s (23.75ns/i, 68clocks/i)
Calculating -------------------------------------
                         before       after
         :sym.to_sym    60.844M     69.503M i/s -    126.321M times in 2.076139s 1.817497s

Comparison:
                      :sym.to_sym
               after:  69502715.5 i/s
              before:  60844185.9 i/s - 1.14x  slower
```

### MJIT
MJIT has method inlining, so calling this method becomes faster. (Using `--runner=block` because MJIT doesn't compile top-level ISeqs. So the i/s numbers can't be compared with other results)

```
$ benchmark-driver -v --runner=block --chruby 'before --mjit;after --mjit' ':sym.to_sym'
before --mjit: ruby 3.2.0dev (2022-11-15T08:34:00Z master 1a9e87fe3a) +MJIT [x86_64-linux]
after --mjit: ruby 3.2.0dev (2022-11-15T08:38:30Z builtin-sym-to-sym 43fb0e94cb) +MJIT [x86_64-linux]
Warming up --------------------------------------
         :sym.to_sym       29.554M i/s -     29.941M times in 1.013091s (33.84ns/i, 126clocks/i)
Calculating -------------------------------------
                     before --mjit  after --mjit
         :sym.to_sym       37.037M       47.251M i/s -     88.661M times in 2.393876s 1.876395s

Comparison:
                      :sym.to_sym
        after --mjit:  47250836.8 i/s
       before --mjit:  37036677.0 i/s - 1.28x  slower
```

### YJIT
YJIT doesn't have method inlining yet, but this PR doesn't make YJIT slower.

```
$ benchmark-driver -v --chruby 'before --yjit;after --yjit' ':sym.to_sym'
before --yjit: ruby 3.2.0dev (2022-11-15T08:34:00Z master 1a9e87fe3a) +YJIT [x86_64-linux]
after --yjit: ruby 3.2.0dev (2022-11-15T08:38:30Z builtin-sym-to-sym 43fb0e94cb) +YJIT [x86_64-linux]
Warming up --------------------------------------
         :sym.to_sym       44.355M i/s -     44.893M times in 1.012133s (22.55ns/i, 85clocks/i)
Calculating -------------------------------------
                     before --yjit  after --yjit
         :sym.to_sym       60.814M       67.980M i/s -    133.066M times in 2.188084s 1.957411s

Comparison:
                      :sym.to_sym
        after --yjit:  67980378.9 i/s
       before --yjit:  60813714.4 i/s - 1.12x  slower
```